Created an object "CustomMap" that manages the map
